scholar-search/requirements.txt
# Testing and code checking
hypothesis
pytest
python-ta~=2.9.1

# Project 2 Specific
Flask~=3.1.0
requests~=2.32.3
kaggle~=1.7.4.5
gunicorn~=23.0.0
python-dotenv~=1.1.0
